fix tests post-refactoring
